+++ src/main/java/com/ipfms/domain/model/CustomAttributeValue.java
@@ -11,43 +11,43 @@ public class CustomAttributeValue {
     @Id
     @GeneratedValue(strategy = GenerationType.AUTO)
     @JsonApiId
-    private long id;
-    private long attrId;
-    private long recordId;
+    private Integer id;
+    private Integer attrId;
+    private Integer recordId;
     private String value;
 
     public CustomAttributeValue() {
         super();
     }
 
-    public CustomAttributeValue(long id, long attrId, long recordId, String value) {
+    public CustomAttributeValue(Integer id, Integer attrId, Integer recordId, String value) {
         this.id = id;
         this.attrId = attrId;
         this.recordId = recordId;
         this.value = value;
     }
 
-    public long getId() {
+    public Integer getId() {
         return id;
     }
 
-    public void setId(long id) {
+    public void setId(Integer id) {
         this.id = id;
     }
 
-    public long getAttrId() {
+    public Integer getAttrId() {
         return attrId;
     }
 
-    public void setAttrId(long attrId) {
+    public void setAttrId(Integer attrId) {
         this.attrId = attrId;
     }
 
-    public long getRecordId() {
+    public Integer getRecordId() {
         return recordId;
     }
 
-    public void setRecordId(long recordId) {
+    public void setRecordId(Integer recordId) {
         this.recordId = recordId;
     }
