+++ src/main/java/com/ipfms/domain/repository/CustomAttributeRepositoryImpl.java
@@ -8,8 +8,8 @@ import io.katharsis.queryspec.QuerySpec;
 import io.katharsis.repository.ResourceRepositoryBase;
 import io.katharsis.resource.list.ResourceList;
 
-public class CustomAttributeRepositoryImpl extends ResourceRepositoryBase<CustomAttribute, Long>{
-    private Map<Long, CustomAttribute> customAttributes = new HashMap<>();
+public class CustomAttributeRepositoryImpl extends ResourceRepositoryBase<CustomAttribute, Integer>{
+    private Map<Integer, CustomAttribute> customAttributes = new HashMap<>();
     public CustomAttributeRepositoryImpl() {
         super(CustomAttribute.class);
     }
